tests: move taas code from rkt_auth_test.go to rkt_tests.go
